Pull config into it's own module so slot handling logic can be separated
